annualised_heating_system_cost

Current equivalent annualised cost of a home heating system installed at installation_time.

annualised_capex(
  tech,
  kW,
  installation_time,
  installation_type,
  house_type,
  construction_year,
  grant_type,
  params
)

Arguments

tech

heating technology

kW

system capacity

installation_time

time of installation - different from params$yeartime

installation_type

"new" or "swap"

house_type

seai house type

construction_year

year, integer

grant_type

grant type

params

time of interest or current time

Value

euro amount

Details

this function uses a capital reduction factor with with a technology specific Weibull survival function. The annualised capex falls when the system age exceeds the expected lifetime. This corresponds to the intuition that older systems that continue to operate are "free".

The discount rate set from calibration (params$r.).

A present bias (above a capex threshold) params$beta. is included.

Examples

 sapply(2010:2040, function(y) annualised_capex("heat_pump",18,2010,"swap","detached",2003,"WarmerHomes",scenario_params(sD,y)))
#>  [1] 1501.44072 1503.36350 1503.32632 1500.02491 1492.18340 1478.58783
#>  [7] 1459.41728 1433.63655 1399.15003 1355.43337 1302.27839 1239.83660
#> [13] 1168.64685 1089.64169 1004.12812  913.74090  818.96993  722.35597
#> [19]  627.53670  536.48222  450.94521  372.36462  301.79586  239.87363
#> [25]  186.81007  142.42653  106.21325   77.40840   55.08631   38.24459
#> [31]   25.88163
params <- scenario_params(sD,2026)
replicate(100,annualised_capex("heat_pump",13,params$yeartime,"new","detached",2003,"None",params)) %>% table()
#> .
#> 1715.80408497979 
#>              100 
params$beta. <- 0.5
annualised_capex("gas",8,params$yeartime,"new","detached",2003,"None",params)
#> [1] 285.5938
annualised_capex("heat_pump",8,params$yeartime,"new","detached",2003,"BetterEnergyHomes",params)
#> [1] 861.2664