heat_pump_savings

find the financially optimum household heating system replacement. Choose between current system and installing a heat pump.

heat_pump_savings(
  hli,
  tech_old,
  installation_time,
  house_type,
  storeys,
  construction_year,
  region,
  floor_area,
  params
)

Arguments

hli

current HLI

tech_old

current tech

installation_time

year of installation of old tech

house_type

seai house type

storeys

1 or 2+

construction_year

integer

region

region

floor_area

total floor area (m2)

params

current parameter values

Value

one row dataframe

Details

The "savings" are relative to an replacing the current system with a new system with the same technology e.g. gas -> gas

If current system is a heat pump, choose between a heat pump swap or reversion to gas.

This function is used in update_agents to determine the system choice when current heating system fails

Examples

params <- scenario_params(sD,2025)
heat_pump_savings(2,"gas",2010,"detached",2,1995,"Munster",120,params)
#> # A tibble: 1 × 7
#>   tech_cost_stick tech_cost_switch eac_stick eac_switch savings grant_type      
#>             <dbl>            <dbl>     <dbl>      <dbl>   <dbl> <chr>           
#> 1            2157            13863      2390       3136   0.312 BetterEnergyHom…
#> # ℹ 1 more variable: heat_pump_grant <dbl>
params$beta. <- 1
heat_pump_savings(2,"solid_fuel",2010,"detached",2,1995,"Munster",100,params)
#> # A tibble: 1 × 7
#>   tech_cost_stick tech_cost_switch eac_stick eac_switch savings grant_type      
#>             <dbl>            <dbl>     <dbl>      <dbl>   <dbl> <chr>           
#> 1            2480            12655      1836       2266   0.234 BetterEnergyHom…
#> # ℹ 1 more variable: heat_pump_grant <dbl>

heat_pump_savings(2,"heat_pump",2010,"detached",2,1995,"Munster",120,params)
#> # A tibble: 1 × 7
#>   tech_cost_stick tech_cost_switch eac_stick eac_switch savings grant_type      
#>             <dbl>            <dbl>     <dbl>      <dbl>   <dbl> <chr>           
#> 1           10778             2756      2528       2235  -0.116 BetterEnergyHom…
#> # ℹ 1 more variable: heat_pump_grant <dbl>