heat_pump_upgrade_savings

The percentage annual equivalent savings achieved by adopting a heat pump as part of an home energy upgrade.

heat_pump_upgrade_savings(
  hli_old,
  tech_old,
  installation_time,
  house_type,
  storeys,
  construction_year,
  region,
  floor_area,
  cost_model = "logistic",
  params,
  is_fuel_allowance,
  include_grants
)

Arguments

hli_old

starting hli

tech_old

old tech

installation_time

time of installation of old tech

house_type

seai house type

storeys

1 or 2+

construction_year

house construction year

region

region

floor_area

floor area in m2

cost_model

logistic, power or esri

params

scenario params at current yeartime

is_fuel_allowance

default FALSE

include_grants

TRUE or FALSE

Value

a single row dataframe

Details

Savings are expressed in two was (1) relative to the current system and (2) the current technology. Savings are also expressed relative to current system with no upgrades (used in update_agents to determine whether an upgrade is undertaken) If the relative gain is sufficient, a heat pump may be adopted.

For WarmerHomes, it is assumed that pre-2025 retrofits "stick" to current heating technology. Post-2025 the cost optimal choice is made.

Examples

params <- scenario_params(sD,2026)
heat_pump_upgrade_savings(2.5,"gas",2000,"semi_detached",2,1990,"Munster",106,"logistic",params,is_fuel_allowance=FALSE,include_grants=TRUE)
#> # A tibble: 1 × 16
#>   hli_stick eac_stick upgrade_cost_stick tech_cost_stick grant_type_stick 
#>       <dbl>     <dbl>              <dbl>           <dbl> <chr>            
#> 1      2.50      2607                 90            2226 BetterEnergyHomes
#> # ℹ 11 more variables: upgrade_grant_stick <dbl>, heat_pump_grant_stick <dbl>,
#> #   hli_switch <dbl>, eac_switch <dbl>, upgrade_cost_switch <dbl>,
#> #   tech_cost_switch <dbl>, grant_type_switch <chr>,
#> #   upgrade_grant_switch <dbl>, heat_pump_grant_switch <dbl>, savings <dbl>,
#> #   eac_old <dbl>

heat_pump_upgrade_savings(1.1,"gas",2000,"semi_detached",2,1990,"Munster",106,"logistic",params,is_fuel_allowance=TRUE,include_grants=TRUE)
#> # A tibble: 1 × 16
#>   hli_stick eac_stick upgrade_cost_stick tech_cost_stick grant_type_stick 
#>       <dbl>     <dbl>              <dbl>           <dbl> <chr>            
#> 1      1.10      1382                 36            1848 BetterEnergyHomes
#> # ℹ 11 more variables: upgrade_grant_stick <dbl>, heat_pump_grant_stick <dbl>,
#> #   hli_switch <dbl>, eac_switch <dbl>, upgrade_cost_switch <dbl>,
#> #   tech_cost_switch <dbl>, grant_type_switch <chr>,
#> #   upgrade_grant_switch <dbl>, heat_pump_grant_switch <dbl>, savings <dbl>,
#> #   eac_old <dbl>