optimise_heat

Find the financially optimum household heating system replacement with no BER upgrade. The choice is between replacing current system or installing a heat pump.

optimise_heat(
  hli,
  tech_old,
  installation_time,
  house_type,
  storeys,
  construction_year,
  region,
  floor_area,
  params,
  include_rebound = FALSE
)

Arguments

hli

the current W/Km2 not upgraded

tech_old

new ber kWh/m2/year

installation_time

year of installation of old tech

house_type

seai house type

storeys

1 or 2+

construction_year

integer

region

region

floor_area

total floor area (m2)

params

current parameter values

include_rebound

default FALSE

Value

data frame (2 or 1 rows)

Details

The function is used for example in the case of a heating system failure, where a household does not consider heat loss efficiency upgrades.

If current system is a heat pump, choose between a heat pump swap or reversion to gas.

Examples

params <- scenario_params(sD,2026)
optimise_heat(2.6,"oil",2000,"terraced",2,1980,"Munster",90,params)
#> # A tibble: 2 × 10
#>   tech_old tech_new  new_cost old_cost savings upgrade_cost heating_sys_cost
#>   <chr>    <chr>        <dbl>    <dbl>   <dbl>        <dbl>            <dbl>
#> 1 oil      oil          1961.    2596.  -0.244            0            2361.
#> 2 oil      heat_pump    3396.    2596.   0.308            0           13710.
#> # ℹ 3 more variables: grant_type <chr>, upgrade_grant <dbl>,
#> #   heat_pump_grant <dbl>

params <- scenario_params(sD,2026)
optimise_heat(4,"oil",2000,"terraced",2,1980,"Munster",90,params)
#> # A tibble: 2 × 10
#>   tech_old tech_new  new_cost old_cost savings upgrade_cost heating_sys_cost
#>   <chr>    <chr>        <dbl>    <dbl>   <dbl>        <dbl>            <dbl>
#> 1 oil      oil          2868.    3906.  -0.266    -2.12e-12            2792.
#> 2 oil      heat_pump    4366.    3906.   0.118     0                  17516.
#> # ℹ 3 more variables: grant_type <chr>, upgrade_grant <dbl>,
#> #   heat_pump_grant <dbl>
optimise_heat(5.4,"oil",2000,"terraced",2,1980,"Munster",90,params)
#> # A tibble: 2 × 10
#>   tech_old tech_new  new_cost old_cost savings upgrade_cost heating_sys_cost
#>   <chr>    <chr>        <dbl>    <dbl>   <dbl>        <dbl>            <dbl>
#> 1 oil      oil          3778.    5216. -0.276             0            3223.
#> 2 oil      heat_pump    5609.    5216.  0.0753            0           21323.
#> # ℹ 3 more variables: grant_type <chr>, upgrade_grant <dbl>,
#> #   heat_pump_grant <dbl>