optimise_upgrade

optimise_upgrade finds the financially optimum household energy efficiency upgrade (fabric and replacement heating technology)

Usage

optimise_upgrade(
  hli_old,
  tech_old,
  installation_time,
  house_type,
  storeys,
  construction_year,
  region,
  floor_area,
  cost_model = "logistic",
  params,
  upgrade_heat = TRUE,
  is_fuel_allowance = FALSE,
  include_grants = TRUE,
  include_rebound = FALSE
)

Arguments

hli_old

heat loss indicator before upgrade W/Km2

tech_old

current heating technology

installation_time

year of installation of old (current) tech

house_type

seai house type (detached, semi_detached, terraced, apartment)

storeys

number of storeys 1, 2, 3

construction_year

year of construction integer

region

region (Dublin, Munster, Rest of Leinster, Ulster/Connacht)

floor_area

treated floor area (m2)

cost_model

efficiency upgrade cost model - only "logistic" at the moment

params

current parameter values from scenario_params(sD, yeartime)

upgrade_heat

if TRUE then upgrade the current heating system

is_fuel_allowance

TRUE/FALSE

include_grants

defaults to TRUE

include_rebound

defaults to FALSE

Value

a two row dataframe. The new EAC "new_cost"

Details

The potential new heating technologies (tech_new) considered are current tech and a heat pump. Choices such as gas to oil are not considered,

If fuel_allowance is TRUE and the household qualifies for WarmerHomes. It is assumed that the property is upgraded to B2 (120 kWh/m2/y) standard.

All grant_types apart from WarmerHomes requires BER optimisation.

Returns a 2 row dataframe with tech_new the existing tech or a heat pump. If the existing tech is a heat pump, reversion to gas is also evaluated.

Examples


params <- scenario_params(sD,2026)

optimise_upgrade(4,"gas",2010,"detached",2,2003,"Dublin",100,"logistic",params,TRUE,FALSE,TRUE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>   <dbl>
#> 1 gas      gas             4    2.96    319.    229.    2858.    3556.  -0.196
#> 2 gas      heat_pump       4    2.30    319.    111.    2634.    3556.  -0.259
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>

optimise_upgrade(3,"gas",2010,"detached",2,2010,"Dublin",100,"logistic",params,TRUE,FALSE,TRUE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>   <dbl>
#> 1 gas      gas             3    2.72    249.    213.    2473.    2707. -0.0864
#> 2 gas      heat_pump       3    2.27    249.    110.    2240.    2707. -0.173 
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>

optimise_upgrade(2.4,"gas",2010,"detached",2,2010,"Dublin",175,"logistic",params,TRUE,FALSE,TRUE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost  savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>    <dbl>
#> 1 gas      gas           2.4    2.31    207.    187.    3522.    3726. -0.0546 
#> 2 gas      heat_pump     2.4    2.08    207.    104.    3698.    3726. -0.00748
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>

optimise_upgrade(2.9,"gas", 2015, "detached",2, 2003, "Munster", 100,"logistic",params,upgrade_heat=TRUE,is_fuel_allowance=FALSE,include_rebound=FALSE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>   <dbl>
#> 1 gas      gas           2.9    2.61    226.    206.    2380.    2509. -0.0513
#> 2 gas      heat_pump     2.9    2.12    226.    105.    1848.    2509. -0.263 
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>

optimise_upgrade(2.9,"gas", 2015, "detached",2, 2003, "Munster", 200,"logistic",params,upgrade_heat=TRUE,is_fuel_allowance=FALSE,include_rebound=FALSE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>   <dbl>
#> 1 gas      gas           2.9    2.61    226.    206.    4588.    4804. -0.0449
#> 2 gas      heat_pump     2.9    2.27    226.    110.    4529.    4804. -0.0573
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>

optimise_upgrade(3,"gas",2000,"detached",2,1990,"Dublin",120,"logistic",params,upgrade_heat=TRUE,is_fuel_allowance=FALSE,include_grants=TRUE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>   <dbl>
#> 1 gas      gas             3    2.75    328.    215.    2935.    4305.  -0.318
#> 2 gas      heat_pump       3    2.27    328.    110.    2824.    4305.  -0.344
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>

#WarmerHomes

optimise_upgrade(4,"gas", 2015, "detached",2, 2003, "Munster", 100,"logistic",params,upgrade_heat=TRUE,is_fuel_allowance=TRUE,include_rebound=FALSE)
#> # A tibble: 2 × 14
#>   tech_old tech_new  hli_old hli_new ber_old ber_new new_cost old_cost  savings
#>   <chr>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>    <dbl>
#> 1 gas      gas             4    1.27    297.     120    1313.    3379. -0.612  
#> 2 gas      heat_pump       4    2.59    297.     120    3372.    3379. -0.00213
#> # ℹ 5 more variables: upgrade_cost <dbl>, heating_sys_cost <dbl>,
#> #   grant_type <chr>, upgrade_grant <dbl>, heat_pump_grant <dbl>