retrofit_cost_model_logistic

Building fabric + ventilation retrofit upgrade cost bute excluding tech such as solar PV or Heat Pumps

retrofit_cost_model_logistic(
  hli_old,
  hli_new,
  house_type,
  storeys,
  region = "Dublin",
  floor_area = 100,
  params
)

Arguments

hli_old

hli before upgrade

hli_new

hli after upgrade

house_type

house type

storeys

number of storeys

region

region

floor_area

floor area (m2)

params

current parameters

Value

euros

Details

The default and possibly the most plausible cost model, based on HLI to exclude effects of heating technology efficiency.

Economies of scale currently use a \(area^\frac{2}{3}\) scaling law.

Examples

params <- scenario_params(sD,2018)
retrofit_cost_model_logistic(5,4,"semi_detached",2,"Dublin",100,params)
#> [1] 4795.228

params <- scenario_params(sD,2024)
retrofit_cost_model_logistic(4.5,2.0,"detached",2,"Munster",100,params)
#> [1] 34861.98
retrofit_cost_model_logistic(2.41,1.91,"detached",1,"Dublin",175,params) #UCD example
#> [1] 39700.22