runABM

Runs home energy efficiency system adoption simulation on artificial society of ~792 agents. Each run is performed on an independently generated social network with randomisation from initialise_agents() and with a random micro_calibration run (index 1..100)

runABM(
  sD,
  Nrun = 1,
  simulation_end = 2030,
  resample_society = F,
  n_unused_cores = 2,
  use_parallel = T,
  ignore_social = F,
  quiet = TRUE
)

Arguments

sD

scenario set-up dataframe, typically read with readr::read_xlxs(...,sheet=scenario)

Nrun

integer, number runs

simulation_end

the final year of simulation of early termination is required

resample_society

if TRUE resample hp_society_oo with replacement to capture additional variability

n_unused_cores

number of cores left unused in parallel/foreach. Recommended values 2 or 1.

use_parallel

if TRUE uses multiple cores. Use FALSE for diagnostic runs on a single core.

ignore_social

if TRUE ignore social network effects. Default is FALSE

quiet

if TRUE messaging is reduced

Value

a three component list - simulation output, scenario setup, meta-parameters

Details

Bi-monthly timesteps.

Good luck.