heating_upgrade_tensor

This is a core function of hpmicrosimr. heating_upgrade_tensor calculates an element of the heating system upgrade cost tensor i.e. the costs and cost savings (hli_old,tech_old) -> (hli_new,tech_new)

heating_upgrade_tensor(
  hli_old,
  hli_new,
  tech_old,
  installation_time,
  tech_new,
  house_type,
  storeys,
  construction_year,
  region,
  floor_area,
  cost_model = "logistic",
  params,
  upgrade_heat = FALSE,
  is_fuel_allowance = FALSE,
  include_grants = TRUE,
  include_rebound = FALSE
)

Arguments

hli_old

old ber double

hli_new

new ber double

tech_old

old tech

installation_time

time of installation of current system (for estimate of old operating cost only)

tech_new

new tech

house_type

house type in seai codes

storeys

number of storeys 1, 2+

construction_year

year,integer

region

region

floor_area

floor area of property in m2

cost_model

ber upgrade cost model (power, logistic or esri) default logistic

params

params

upgrade_heat

TRUE if current heating system is upgraded

is_fuel_allowance

TRUE/FALSE

include_grants

TRUE/FALSE

include_rebound

defulat FALSE

Value

list

Details

fabric upgrade costs are discounted with infinite lifetime. Present bias (params$beta.) is included for new investments. An additional sludge/

the new tech is installed at current time (params$yeartime). The old tech installation time is also specified.

All applicable grants used by default, the effects of grants can optionally be excluded by setting include_grants=FALSE

Optionally, the current (old) heating system can be retained

Used with optimise_upgrade to determine the optimal upgrade path tech_old -> tech_new and hli_old -> hli_new

Examples


params <- scenario_params(sD,2026)

heating_upgrade_tensor(1.8,1.5,"gas",2005,"heat_pump","semi_detached",2,1990,"Dublin",100,cost_model="logistic",params,TRUE,FALSE,TRUE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>     
#> 1    1272.    2226.  -0.428       14954.           11172. OSS       
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>

heating_upgrade_tensor(4,3,"oil",2000,"heat_pump","terraced",2,1980,"Munster",90,"logistic",params,TRUE,FALSE,TRUE,include_rebound=TRUE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type       
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>            
#> 1    3580.    2789.   0.284        4795.           14797. BetterEnergyHomes
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>

heating_upgrade_tensor(2,1.2,"oil",2000,"heat_pump","semi_detached",2, 1980,"Munster",100,"logistic",params,TRUE,FALSE,TRUE,include_rebound=FALSE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>     
#> 1    1755.    2243.  -0.217       31990.           10265. OSS       
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>

heating_upgrade_tensor(2,1.2,"gas",2000,"gas","detached",2, 1980,"Dublin",100,"logistic",params,TRUE,FALSE,TRUE,include_rebound=FALSE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>     
#> 1    2614.    2450.  0.0670       47447.            1857. OSS       
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>
heating_upgrade_tensor(3,2,"gas",2000,"gas","detached",2, 1980,"Dublin",100,"logistic",params,TRUE,FALSE,TRUE,include_rebound=FALSE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type       
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>            
#> 1    3213.    3610.  -0.110       32179.            2061. BetterEnergyHomes
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>
heating_upgrade_tensor(3,1,"gas",2000,"gas","detached",2, 1980,"Dublin",100,"logistic",params,TRUE,FALSE,TRUE,include_rebound=FALSE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>     
#> 1    4708.    3610.   0.304       92514.            1806. OSS       
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>
heating_upgrade_tensor(3,2,"oil",2000,"heat_pump","semi_detached",2, 1980,"Munster",100,"logistic",params,TRUE,FALSE,TRUE,include_rebound=FALSE)
#> # A tibble: 1 × 8
#>   new_cost old_cost savings upgrade_cost heating_sys_cost grant_type
#>      <dbl>    <dbl>   <dbl>        <dbl>            <dbl> <chr>     
#> 1    1887.    3282.  -0.425       20213.           12682. OSS       
#> # ℹ 2 more variables: upgrade_grant <dbl>, heat_pump_grant <dbl>
ber_from_hli(1.27,"gas",params$yeartime,params)
#> [1] 119.8083