update_agents

micro-simulation time-step updater

update_agents(
  sD,
  yeartime,
  agents_in,
  social_network,
  ignore_social = F,
  cal_run = 50,
  quiet = TRUE
)

Arguments

sD

scenario dataframe

yeartime

decimal time

agents_in

input agent dataframe

social_network

artifical social network

ignore_social

option to ignore social effects. Default is FALSE.

cal_run

microcalibration run index between 1 and 100

quiet

TRUE to suppress messages

Value

updated agent dataframe

Details

The workhorse ABM function.Within a scenario, does a single month update of the agent characteristics. A random sample of agents evaluates their economic and social utilities. If these exceed their individual threshold an optimal PV_BESS system is adopted.

Examples


agents_in <- initialise_agents(sD,2015,100)
#> Joining with `by = join_by(q6)`
#> Joining with `by = join_by(qc2)`
#> Joining with `by = join_by(q1)`
#> Joining with `by = join_by(serial)`
#> Joining with `by = join_by(qh)`
social_network <- make_artificial_society(hp_society_oo %>% dplyr::filter(serial %in% agents_in$serial),homophily)
agents_1 <- update_agents(sD,2015+1/6,agents_in,social_network,cal_run=10,quiet=FALSE)
#> # A tibble: 0 × 30
#> # ℹ 30 variables: serial <chr>, w_q52 <dbl>, w_q13 <dbl>, w_theta <dbl>,
#> #   region <chr>, house_type <chr>, storeys <int>, construction_year <dbl>,
#> #   ber <dbl>, hli <dbl>, floor_area <dbl>, tech <chr>,
#> #   heating_install_time <dbl>, income <dbl>, fuel_allowance <lgl>, kW <dbl>,
#> #   q52 <dbl>, eac <dbl>, eac_actual <dbl>, upgrade <lgl>, failure <lgl>,
#> #   upgrade_cost <dbl>, grant_type <chr>, tech_cost <dbl>, upgrade_grant <dbl>,
#> #   eac_stick <dbl>, eac_switch <dbl>, savings <dbl>, hp_grant_type <chr>, …
#> [1] "time 2015.2 number of system breakdowns 6"
#> [1] "time 2015.2 number of efficiency upgrades 2"
#> [1] "time 2015.2 number of heat pump adopters following system breakdown 0"
#> [1] "number of heat pump adopters as part of efficiency upgrade 0"
agents_2 <- update_agents(sD,2015+1/3,agents_1,social_network,cal_run=10,quiet=FALSE)
#> # A tibble: 0 × 31
#> # ℹ 31 variables: serial <chr>, w_q52 <dbl>, w_q13 <dbl>, w_theta <dbl>,
#> #   region <chr>, house_type <chr>, storeys <int>, construction_year <dbl>,
#> #   ber <dbl>, hli <dbl>, floor_area <dbl>, tech <chr>,
#> #   heating_install_time <dbl>, income <dbl>, fuel_allowance <lgl>, kW <dbl>,
#> #   q52 <dbl>, eac <dbl>, eac_actual <dbl>, upgrade <lgl>, failure <lgl>,
#> #   upgrade_cost <dbl>, grant_type <chr>, tech_cost <dbl>, upgrade_grant <dbl>,
#> #   heating_sys_cost <dbl>, eac_stick <dbl>, eac_switch <dbl>, savings <dbl>, …
#> [1] "time 2015.3 number of system breakdowns 8"
#> [1] "time 2015.3 number of efficiency upgrades 1"
#> [1] "time 2015.3 number of heat pump adopters following system breakdown 0"
#> [1] "number of heat pump adopters as part of efficiency upgrade 0"